requirements.txt
torch>=2.0
torchvision>=0.15
transformers>=4.40
Pillow>=9.0
tqdm>=4.65.0